Revert "#17774 #17392 Move SQLPartitionScanner to the model package"
