+++ plugins/org.jkiss.dbeaver.model.sql/META-INF/MANIFEST.MF
@@ -35,7 +35,6 @@ Require-Bundle: org.eclipse.equinox.security,
  org.eclipse.core.expressions,
  org.eclipse.core.commands,
  org.eclipse.core.resources,
- org.eclipse.jface.text,
  com.google.gson,
  org.eclipse.text;visibility:=reexport,
  org.jkiss.dbeaver.model;visibility:=reexport
